// src/app/core/models/settings/department.ts
export class Department {
    departmentID: number;
    departmentName: string;
    departmentNameBangla: string;
    shortName:string;
    isActive: boolean;
    organizationID: number;
    orderNo: number | null;
    parentDepartmentID: number;
    parentDepartmentName: string;
    parentDepartmentNameBangla: string;
    organizationName: string;
    organizationNameBangla: string;
}