// src/app/core/models/settings/departmentDesginationMap.ts
export class DepartmentDesginationMap {
    departmentDesignationMapID: number;
    departmentID: number;
    designationID: number;
    createdBy: number;
    createdDate: string;
    departmentName: string;
    designationName: string;
    departmentNameBangla: string;
    designationNameBangla: string;
}